Don't detach clients automatically
I forgot I even had this. This was why I was having issues reusing my sessions.
My normal workflow was to have the session open on both my desktop and while
SSHing into it, but WFH has me with remote-only sessions. I'll just have to
remember to always attach to the session.